Separated manager folder and clean structer
