+++ client/src/app/app.routes.ts
@@ -24,7 +24,7 @@ import { secretaryGuard } from './guards/secretary.guard';
 import { studentGuard } from './guards/student.guard';
 import { teacherGuard } from './guards/teacher.guard';
 import { RecoveryComponent } from './components/recovery/recovery.component';
-import { ManagerPanelComponent } from './components/manager-panel/manager-panel.component';
+import { ManagerPanelComponent } from './components/manager/manager-panel/manager-panel.component';
 import { CourseCreateComponent } from './components/courses/pages/create/course-create.component';
 import { CourseCardComponent } from './components/courses/course-card/course-card.component';
 import { CourseEditComponent } from './components/courses/pages/edit/course-edit.component';
