+++ client/src/app/components/manager/manager-panel/manager-panel.component.html
