+++ client/src/app/components/manager/register/register-secretary/register-secretary.component.ts
@@ -6,12 +6,12 @@ import { MatRadioModule } from '@angular/material/radio';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
-import { AutoFocusDirective } from '../../../directives/auto-focus.directive';
-import { ManagerService } from '../../../services/manager.service';
-import { RegisterUser } from '../../../models/register-user.model';
+import { AutoFocusDirective } from '../../../../directives/auto-focus.directive';
+import { ManagerService } from '../../../../services/manager.service';
+import { RegisterUser } from '../../../../models/register-user.model';
 import { MatIconModule } from "@angular/material/icon";
 import moment, { Moment } from 'moment-jalaali';
-import { DatepickerComponent } from '../../../datepicker/datepicker.component';
+import { DatepickerComponent } from '../../../../datepicker/datepicker.component';
 
 @Component({
   selector: 'app-register-secretary',
