+++ client/src/app/components/manager/register/register-teacher/register-teacher.component.scss
