+++ client/src/app/components/manager/register/shared-form-styles.scss
